Non-blocking synchronous APIs

+++ microservices/comment-service/src/main/java/se/magnus/microservices/core/comment/services/CommentServiceImpl.java
@@ -1,6 +1,6 @@
 package se.magnus.microservices.core.comment.services;
 
-import java.util.ArrayList;
+import org.reactivestreams.Publisher;
 import java.util.List;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -9,6 +9,8 @@ import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.dao.DuplicateKeyException;
 import org.springframework.web.bind.annotation.RestController;
 
+import reactor.core.publisher.Flux;
+import reactor.core.scheduler.Scheduler;
 import se.magnus.api.core.comment.Comment;
 import se.magnus.api.core.comment.CommentService;
 import se.magnus.api.core.ingredient.Ingredient;
@@ -17,6 +19,9 @@ import se.magnus.microservices.core.comment.persistence.CommentRepository;
 import se.magnus.util.exceptions.InvalidInputException;
 import se.magnus.util.exceptions.NotFoundException;
 import se.magnus.util.http.ServiceUtil;
+import java.util.function.Supplier;
+
+import static java.util.logging.Level.FINE;
 
 @RestController
 public class CommentServiceImpl implements CommentService {
@@ -27,17 +32,26 @@ public class CommentServiceImpl implements CommentService {
 	private final CommentMapper mapper;
 	private final CommentRepository repository;
 
+	private final Scheduler scheduler;
+
 	@Autowired
-	public CommentServiceImpl(ServiceUtil serviceUtil, CommentMapper mapper, CommentRepository repository) {
+	public CommentServiceImpl(Scheduler scheduler, ServiceUtil serviceUtil, CommentMapper mapper, CommentRepository repository) {
+		this.scheduler = scheduler;
 		this.serviceUtil = serviceUtil;
 		this.mapper = mapper;
 		this.repository = repository;
 	}
 
 	@Override
-	public List<Comment> getComments(int mealId) {
+	public Flux<Comment> getComments(int mealId) {
 		if (mealId < 1)
 			throw new InvalidInputException("Invalid mealId: " + mealId);
+
+		LOG.info("Will get comments for meal with id={}", mealId);
+
+		return asyncFlux(() -> Flux.fromIterable(getByMealId(mealId))).log(null, FINE);
+	}
+	protected List<Comment> getByMealId(int mealId) {
 
 		List<CommentEntity> entityList = repository.findByMealId(mealId);
 		List<Comment> list = mapper.entityListToApiList(entityList);
@@ -50,6 +64,8 @@ public class CommentServiceImpl implements CommentService {
 
 	@Override
 	public Comment createComment(Comment body) {
+		if (body.getMealId() < 1) throw new InvalidInputException("Invalid mealId: " + body.getMealId());
+
 		try {
 			CommentEntity entity = mapper.apiToEntity(body);
 			CommentEntity newEntity = repository.save(entity);
@@ -64,8 +80,14 @@ public class CommentServiceImpl implements CommentService {
 
 	@Override
 	public void deleteComments(int mealId) {
+		if (mealId < 1) throw new InvalidInputException("Invalid mealId: " + mealId);
+
 		LOG.debug("deleteComments: tries to delete comment for the meal with mealId: {}", mealId);
         repository.deleteAll(repository.findByMealId(mealId));
+	}
+
+	private <T> Flux<T> asyncFlux(Supplier<Publisher<T>> publisherSupplier) {
+		return Flux.defer(publisherSupplier).subscribeOn(scheduler);
 	}
 
 }
